Refactoring : simplify structure by removing the python subdirectory and moving its content in parent
