+++ src/simple_ai/api/grpc/chat/client.py
