+++ src/simple_ai/api/grpc/chat/model.py
