+++ src/simple_ai/api/grpc/chat/server.py
