+++ src/simple_ai/api/grpc/completion/client.py
