+++ src/simple_ai/api/grpc/completion/model.py
