+++ src/simple_ai/api/grpc/completion/server.py
