+++ src/simple_ai/api/grpc/embedding/client.py
