+++ src/simple_ai/api/grpc/embedding/llm_embed_pb2_grpc.py
