+++ src/simple_ai/api/grpc/embedding/model.py
