+++ src/simple_ai/api/grpc/embedding/server.py
